back/requirements.txt
asgiref==3.6.0
django-api==0.1.3
django-rest-framework==0.1.0
Django==4.1.7
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.5.0
pip==21.3.1
pytz==2022.7.1
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.7
wheel==0.37.1
